+++ lib/src/elements/th_line.dart
@@ -6,7 +6,11 @@ part of 'th_element.dart';
 // side of a pitch, higher side of a chimney and interior of a passage are on the left side of
 // pitch, chimney or wall symbols, respectively.
 class THLine extends THElement
-    with THHasOptionsMixin, THIsParentMixin, MPBoundingBox, THHasPLATypeMixin {
+    with
+        THHasOptionsMixin,
+        THIsParentMixin,
+        MPBoundingBoxMixin,
+        THHasPLATypeMixin {
   final THLineType lineType;
 
   List<int>? _lineSegmentMPIDs;